+++ Basil/include/Debug/Instrumentor.h
@@ -54,7 +54,7 @@ namespace Basil
 			bool stopped;
 	};
 
-	#define PROFILE 1
+	#define PROFILE 0
 	#if PROFILE
 		#if defined(__GNUC__) || (defined(__MWERKS__) && (__MWERKS__ >= 0x3000)) || (defined(__ICC) && (__ICC >= 600)) || defined(__ghs__)
 			#define FUNC_SIG __PRETTY_FUNCTION__