+++ Basil/include/Platform/OpenGL/OpenGLRendererAPI.h
@@ -14,6 +14,6 @@ namespace Basil
 			void setViewport(unsigned int x, unsigned int y, unsigned int width, unsigned int height) override;
 			void setClearColor(const glm::vec4& color) override;
 			void clear() override;
-			void drawIndexed(const Shared<VertexArray>& vao) override;
+			void drawIndexed(const Shared<VertexArray>& vao, uint32_t indexCount = 0) override;
 	};
 }