+++ Basil/include/Platform/OpenGL/OpenGLVertexBuffer.h
@@ -15,9 +15,11 @@ namespace Basil
 	{
 		public:
 			OpenGLVertexBuffer(std::vector<float>& vertices);
+			OpenGLVertexBuffer(uint32_t size);
 			~OpenGLVertexBuffer() override;
 			void bind() const override;
 			void unbind() const override;
+			void setData(const void* data, uint32_t size) override;
 			const BufferLayout& getLayout() const override;
 			void setLayout(const BufferLayout& layout) override;
 		private: