+++ Basil/include/Renderer/Renderer2D.h
@@ -16,6 +16,7 @@ namespace Basil
 			static void shutdown();
 			static void beginScene(const OrthographicCamera& camera);
 			static void endScene();
+			static void flush();
 			static void drawQuad(const Transform& transform, const glm::vec4& color);
 			static void drawQuad(const Transform& transform, float textureScale, const Shared<Texture2D>& texture, const glm::vec4& tintColor = glm::vec4(1.0f));
 	};