+++ Basil/include/Renderer/VertexBuffer.h
@@ -15,8 +15,10 @@ namespace Basil
 			virtual ~VertexBuffer() {};
 			virtual void bind() const = 0;
 			virtual void unbind() const = 0;
+			virtual void setData(const void* data, uint32_t size) = 0;
 			virtual const BufferLayout& getLayout() const = 0;
 			virtual void setLayout(const BufferLayout& layout) = 0;
+			static Shared<VertexBuffer> create(uint32_t size);
 			static Shared<VertexBuffer> create(std::vector<float>& vertices);
 	};
 }